+++ crates/rust-client/src/store/note_record/input_note_record/states/consumed_unauthenticated_local.rs
@@ -1,6 +1,6 @@
 use alloc::string::ToString;
 
-use miden_objects::block::BlockHeader;
+use miden_objects::block::{BlockHeader, BlockNumber};
 use miden_objects::note::{NoteId, NoteInclusionProof, NoteMetadata};
 use miden_objects::transaction::TransactionId;
 
@@ -14,7 +14,7 @@ pub struct ConsumedUnauthenticatedLocalNoteState {
     /// information.
     pub metadata: NoteMetadata,
     /// Block height at which the note was nullified.
-    pub nullifier_block_height: u32,
+    pub nullifier_block_height: BlockNumber,
     /// Information about the submission of the note.
     pub submission_data: NoteSubmissionData,
 }
@@ -30,7 +30,7 @@ impl NoteStateHandler for ConsumedUnauthenticatedLocalNoteState {
 
     fn consumed_externally(
         &self,
-        _nullifier_block_height: u32,
+        _nullifier_block_height: BlockNumber,
     ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(None)
     }
@@ -55,7 +55,7 @@ impl NoteStateHandler for ConsumedUnauthenticatedLocalNoteState {
     fn transaction_committed(
         &self,
         _transaction_id: TransactionId,
-        _block_height: u32,
+        _block_height: BlockNumber,
     ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Err(NoteRecordError::InvalidStateTransition(
             "Only processing notes can be committed in a local transaction".to_string(),
@@ -88,7 +88,7 @@ impl miden_tx::utils::Deserializable for ConsumedUnauthenticatedLocalNoteState {
         source: &mut R,
     ) -> Result<Self, miden_tx::utils::DeserializationError> {
         let metadata = NoteMetadata::read_from(source)?;
-        let nullifier_block_height = u32::read_from(source)?;
+        let nullifier_block_height = BlockNumber::read_from(source)?;
         let submission_data = NoteSubmissionData::read_from(source)?;
         Ok(ConsumedUnauthenticatedLocalNoteState {
             metadata,